requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==22.0.0
Pillow==10.4.0
